move devicepointerinfo to its own class
